requirements.txt
asgiref==3.8.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
einops==0.8.1
encodec==0.1.1
filelock==3.18.0
fsspec==2025.5.1
funcy==2.0
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.13.0
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.14.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
